% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{racemap}
\alias{racemap}
\title{Race mapping}
\format{
A named character vector.
}
\usage{
racemap
}
\description{
A named character vector mapping various race categories to a harmonized variable
}
\examples{
racemap
racemap["AI/AN"]
racemap["Black"]
racemap["Doesn't exist in the mapping"]
table(racemap)

}
\keyword{datasets}
