% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_labdata}
\alias{read_labdata}
\title{Read laboratory data}
\usage{
read_labdata(file)
}
\arguments{
\item{file}{A file path}
}
\value{
A data frame
}
\description{
Read laboratory data from a file
}
\examples{
fp <- system.file("extdata/rped.csv", package = "rpkgdemo", mustWork = TRUE)
read_labdata(fp)

}
