% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rped}
\alias{rped}
\title{RPED: R Package Example Data}
\format{
A data frame with 6 columns:
\enumerate{
\item \code{id}: A unique identifier for each individual.
\item \code{sex}: The sex of the individual.
\item \code{age}: The age of the individual.
\item \code{zip}: The ZIP code of the individual's residence.
\item \code{lab1}: A simulated lab value from a Poisson distribution with lambda=2.
\item \code{lab2}: A simulated lab value from a Poisson distribution with lambda=10.
}
}
\usage{
rped
}
\description{
Simulated data for demonstrating the functionality of the rpkgdemo package.
}
\examples{
rped

}
\keyword{datasets}
