% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpkgdemo-package.R
\docType{package}
\name{rpkgdemo-package}
\alias{rpkgdemo}
\alias{rpkgdemo-package}
\title{rpkgdemo: A demo package for building R packages}
\description{
A package to demonstrate building an R package to accompany a workshop on building R packages in Positron. See the getting started vignette for details.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/stephenturner/rpkgdemo}
  \item \url{https://stephenturner.github.io/rpkgdemo/}
  \item Report bugs at \url{https://github.com/stephenturner/rpkgdemo/issues}
}

}
\author{
\strong{Maintainer}: Stephen Turner \email{first.last@example.com} (\href{https://orcid.org/0000-0001-9140-9028}{ORCID})

}
\keyword{internal}
