% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standardize_race}
\alias{standardize_race}
\title{Standardize race variable}
\usage{
standardize_race(x, racecol = "race")
}
\arguments{
\item{x}{A data frame containing a race column to harmonize}

\item{racecol}{The name of the race column. Default is "race".}
}
\value{
A new data frame with an additional \code{.race_standardized} column
}
\description{
Standardize race variable
}
\examples{
set.seed(123)
mydata <- data.frame(reported_race=sample(names(racemap), 20, replace=TRUE))
mydata$reported_race[15] <- "Martian"
mydata
standardize_race(mydata, racecol="reported_race")

}
\seealso{
racemap
}
