% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standardize_sex}
\alias{standardize_sex}
\title{Standardize Sex Variables}
\usage{
standardize_sex(x)
}
\arguments{
\item{x}{A character vector of sex/gender inputs.}
}
\value{
A character vector containing "M", "F", "O", or "U".
}
\description{
Maps various inputs (M, Male, f, Female) to standard HL7/ISO codes (M, F, O, U).
}
\examples{
standardize_sex(c("Male", "f", "Trans female", "nonbinary", "unknown", NA))
}
