% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zipcodes}
\alias{zipcodes}
\title{ZIP code data}
\format{
A data frame with three columns
\enumerate{
\item \code{zip}: The ZIP code.
\item \code{city}: The U.S. city for that zip code.
\item \code{state}: The U.S. state for that zip code.
}
}
\source{
\url{https://postalpro.usps.com/ZIP_Locale_Detail}
}
\usage{
zipcodes
}
\description{
ZIP code data from the US Postal Service.
}
\examples{
zipcodes

}
\keyword{datasets}
